moving functions to utilities
