+++ src/hapi/pipelines/app/__main__.py
@@ -19,6 +19,7 @@ from hdx.utilities.path import temp_dir
 from hapi.pipelines._version import __version__
 from hapi.pipelines.app import load_yamls
 from hapi.pipelines.app.pipelines import Pipelines
+from hapi.pipelines.utilities.process_config_defaults import add_defaults
 
 setup_logging()
 logger = logging.getLogger(__name__)
@@ -62,58 +63,6 @@ def parse_args():
         help="Use saved data",
     )
     return parser.parse_args()
-
-
-def combine_default(country, default):
-    country["input"] = country["input"] + default["input"]
-    country["output"] = country["output"] + default["output"]
-    country["output_hxl"] = country["output_hxl"] + default["output_hxl"]
-    return country
-
-
-def find_defaults(config):
-    default_list = []
-    for key in config:
-        key_components = key.split("_")
-        if len(key_components) == 2 and key_components[1] == "default":
-            default_list.append(key)
-    return default_list
-
-
-def find_matching_top_level_keys(config, default_key):
-    matching_top_level_keys = []
-    default_prefix = default_key.split("_")[0]
-    for top_level_key in config:
-        top_level_key_prefix = top_level_key.split("_")[0]
-        if (
-            top_level_key != default_key
-            and top_level_key_prefix == default_prefix
-        ):
-            matching_top_level_keys.append(top_level_key)
-    return matching_top_level_keys
-
-
-def scraper_add_defaults(config, default, top_level_key):
-    for scraper in default["scrapers_with_defaults"]:
-        if scraper in config[top_level_key]:
-            scraper_config = config[top_level_key][scraper]
-            scraper_config = combine_default(scraper_config, default)
-            config[top_level_key][scraper] = scraper_config
-    return config
-
-
-def add_defaults(config):
-    default_list = find_defaults(config)
-    for default_key in default_list:
-        default = config[default_key]
-        matching_top_level_keys = find_matching_top_level_keys(
-            config, default_key
-        )
-        for top_level_key in matching_top_level_keys:
-            config = scraper_add_defaults(config, default, top_level_key)
-
-        del config[default_key]
-    return config
 
 
 def main(
@@ -211,6 +160,7 @@ if __name__ == "__main__":
     ]
     project_config_dict = load_yamls(project_configs)
     project_config_dict = add_defaults(project_config_dict)
+    print(project_config_dict)
     facade(
         main,
         hdx_key=hdx_key,
